Moved InfluxDB::getNanosecondTimestamp to trait
