+++ src/Drivers/InfluxDB.php
@@ -5,6 +5,7 @@ namespace STS\Metrics\Drivers;
 use InfluxDB\Database;
 use InfluxDB\Point;
 use STS\Metrics\Metric;
+use STS\Traits\ComputesNanosecondTimestamps;
 
 /**
  * Class InfluxDB
@@ -12,6 +13,8 @@ use STS\Metrics\Metric;
  */
 class InfluxDB extends AbstractDriver
 {
+    use ComputesNanosecondTimestamps;
+
     /**
      * @var Database
      */
@@ -82,38 +85,6 @@ class InfluxDB extends AbstractDriver
         $this->points[] = $point;
 
         return $this;
-    }
-
-    /**
-     * A public way tog et the nanosecond precision we desire.
-     *
-     * @param mixed $timestamp
-     *
-     * @return int|null
-     */
-    public function getNanoSecondTimestamp($timestamp = null)
-    {
-        if ($timestamp instanceof \DateTime) {
-            return $timestamp->getTimestamp() * 1000000000;
-        }
-
-        if (strlen($timestamp) == 19) {
-            // Looks like it is already nanosecond precise!
-            return $timestamp;
-        }
-
-        if (strlen($timestamp) == 10) {
-            // This appears to be in seconds
-            return $timestamp * 1000000000;
-        }
-
-        if (preg_match("/\d{10}\.\d{4}/", $timestamp)) {
-            // This looks like a microtime float
-            return (int)($timestamp * 1000000000);
-        }
-
-        // We weren't given a valid timestamp, generate.
-        return (int)(microtime(true) * 1000000000);
     }
 
     /**
