+++ src/Traits/ComputesNanosecondTimestamps.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace STS\Traits;
+
+trait ComputesNanosecondTimestamps
+{
+    /**
+     * A public way tog et the nanosecond precision we desire.
+     *
+     * @param mixed $timestamp
+     *
+     * @return int|null
+     */
+    public function getNanoSecondTimestamp($timestamp = null)
+    {
+        if ($timestamp instanceof \DateTime) {
+            return $timestamp->getTimestamp() * 1000000000;
+        }
+
+        if (strlen($timestamp) == 19) {
+            // Looks like it is already nanosecond precise!
+            return $timestamp;
+        }
+
+        if (strlen($timestamp) == 10) {
+            // This appears to be in seconds
+            return $timestamp * 1000000000;
+        }
+
+        if (preg_match("/\d{10}\.\d{4}/", $timestamp)) {
+            // This looks like a microtime float
+            return (int)($timestamp * 1000000000);
+        }
+
+        // We weren't given a valid timestamp, generate.
+        return (int)(microtime(true) * 1000000000);
+    }
+}
